podział na pliki i funkcje
