+++ php/filter.php
@@ -1,0 +1,8 @@
+<?php
+function filtruj($zmienna) {
+    $zmienna = stripslashes($zmienna); // usuwamy slashe
+
+// usuwamy spacje, tagi html oraz niebezpieczne znaki
+return htmlspecialchars(trim($zmienna));
+}
+?>
