+++ tryb_edycji.php
@@ -1,116 +1,12 @@
 <?php
 include('baza.php');//sprawdzam połączenie
 require_once('polaczeniePDO.php');
+require_once "php/__function_routes.php";
+
 if(!$connect){
     header('location: index.php');//przekierowanie do logowania
 }
 
-function filtruj($zmienna) {
-        $zmienna = stripslashes($zmienna); // usuwamy slashe
-
-// usuwamy spacje, tagi html oraz niebezpieczne znaki
-    return htmlspecialchars(trim($zmienna));
-}
-
-function formularz_dodaj($tabela){
-      echo '<div class="nowe_slowo"><form method="post" action="tryb_edycji.php?zestaw='.$tabela.'&'.$tabela.'=dodaj">
-            Wyrażenie polskie:
-            <input type="text" name="dodaj_element1" style="width: 250px">
-        </br>
-            Wrażenie hiszpańskie:
-            <input type="text" name="dodaj_element2" style="width: 250px">
-        </br>
-            Przykładowe zdanie:
-            <input type="text" name="dodaj_element3" style="width: 250px">
-        </br>
-            <input type="submit" name="submit" value="Wyślij">&nbsp;
-        <input type="reset" value="Wyczyść formularz"></form></br>
-        
-    </div><div class="nowe_slowo">
-    <form method="post" action="tryb_edycji.php?zestaw='.$tabela.'&'.$tabela.'=wyszukaj">
-    Wyszukaj słowo:
-            <input type="text" name="slowo" style="width: 250px">
-        </br>
-            <input type="submit" name="submit" value="Wyślij">&nbsp;
-            <input type="reset" value="Wyczyść formularz"></form>
-    
-    </div>';
-}
-  
-function wyswietlam($tabela, $row){
-$id = $row['id'];
-    
-   echo '<div class="arkusz">';
-    
-   echo '
-         <div class="edytuj">
-         <form method="post" action="tryb_edycji.php?zestaw='.$tabela.'&'.$tabela.'=edytuj&id='.$row['id'].'">
-         <input type="text" name="a'.$id.'" placeholder="Wyrażenie polskie:" value="'.$row['v1'].'">
-         <input type="text" name="b'.$id.'" placeholder="Wrażenie angielskie:"  value="'.$row['v2'].'">
-         <input type="text" name="c'.$id.'" placeholder="Przykładowe zdanie" value="'.$row['zdanie'].'">
-         <input type="submit" name="submit" class="przycisk przycisk1" value="Edytuj!"></form></div>';
-		
-   echo '   
-         <div class="usun"><form method="post" action="tryb_edycji.php?zestaw='.$tabela.'&'.$tabela.'=usun&id='.$row['id'].'">
-         <input type="submit" name="submit" class="przycisk przycisk3" value="Usuń!"></form></div>
-                                                                                    ';
-                                                                                    
-   echo '
-         </div>
-                ';                                                                                    
-}
-
-function wyswietl_nazwe_zestawu($tabela, $pdo){                    
-    try
-    {
-        $querty = "SELECT name_table, flaga FROM info_table WHERE name_table = '$tabela'";
-        $flaga = $pdo ->query($querty) or die('Błąd zapytania SELECT');
-        $row = $flaga->fetch();
-        //echo $tabela. " zmodyfikowano na 0 czyli aktywny";
-        if ($row['flaga'] == 1){
-                echo '
-                   <div class="nowe_slowo">
-                   <div class="select_zestaw">
-                   <div class="wybrany_zestaw_zielony"><h2>'.$tabela.'</h2>
-                     </div>
-                    </div> 
-                   </div>
-                    
-                    ';
-        } else if ($row['flaga'] == 0)   {
-                echo '
-                   <div class="nowe_slowo">
-                   <div class="select_zestaw">
-                   <div class="wybrany_zestaw_czerwony"><h2>'.$tabela.'</h2>
-                     </div>
-                    </div> 
-                   </div>
-                    
-                    ';
-        }      
-        
-       
-    }
-    catch(PDOException $e)
-    {
-       echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
-       echo '</br><a href="tryb_edycji.php?zestaw='.$tabela.'">wróć</a>';
-    }
-                    
-}
-
-function przyciski_aktywuj_deaktywuj($tabela){
-        echo '
-                   <div class="nowe_slowo">
-                   <h2>'.$tabela.'</h2>
-                   
-                    <button id="aktywuj" name="aktywuj" class="przycisk przycisk1">Aktywuj</button>
-                    <button id="deaktywuj" name"deaktywuj" class="przycisk przycisk3">Deaktywuj</button>
-                   </div>
-                    
-                    '; 
-
-} 
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -139,7 +35,7 @@ $tabela = filtruj($_GET['zestaw']);
          echo "<h1>Nie można zrealizować połączenia z żadną bazą danych</h1>";
     }
     
-wyswietl_nazwe_zestawu($tabela, $pdo); 
+load_table_name($_GET['zestaw'], $pdo); 
 
 // menu 
 echo '
@@ -151,7 +47,7 @@ echo '
             <h1>Operacje na słówkach</h1>'
 ;
             przyciski_aktywuj_deaktywuj($tabela);
-            formularz_dodaj($tabela);
+            add_form($tabela);
 echo '
            </div>
    </div>
@@ -172,39 +68,44 @@ if(!isset($_GET[$tabela]))   {
  
  }
  
+function filtruj($zmienna) {
+    $zmienna = stripslashes($zmienna); // usuwamy slashe
+
+// usuwamy spacje, tagi html oraz niebezpieczne znaki
+return htmlspecialchars(trim($zmienna));
+}
+
 // zarzadzanie zestawami
 switch($_GET[$tabela]){
 
 case 'dodaj':          
                    
-        $dodaj_element1 = filtruj($_POST['dodaj_element1']);
-        $dodaj_element2 = filtruj($_POST['dodaj_element2']);
-        $dodaj_element3 = filtruj($_POST['dodaj_element3']); 
+    $dodaj_element1 = filtruj($_POST['dodaj_element1']);
+    $dodaj_element2 = filtruj($_POST['dodaj_element2']);
+    $dodaj_element3 = filtruj($_POST['dodaj_element3']); 
 
-        if (empty($_POST['dodaj_element1']) || empty($_POST['dodaj_element2']))
-        {
-            exit(header("Location: tryb_edycji.php?zestaw=$tabela"));
-        }                
-                                                    
-        //if ($_POST['dodaj_element1'] == $dodaj_element1  && $_POST['dodaj_element2'] == $dodaj_element2 && $_POST['dodaj_element3'] == $dodaj_element3) 
-        //{
-        //exit(header("Location: tryb_edycji.php?zestaw=$tabela"));
-        //}
-        
-        try
-        {
-             $polecenie = "INSERT INTO $tabela ($angielski, $polski, $przyklad , $zdanie, $flaga_baza ) VALUES ('$dodaj_element1', '$dodaj_element2', '1', '$dodaj_element3', '0')";
-             $liczba = $pdo ->exec($polecenie);
-             exit(header("Location: tryb_edycji.php?zestaw=$tabela"));
-        }
-        catch(PDOException $e)
-        {
-           echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
-           echo '</br><a href="tryb_edycji.php?zestaw='.$tabela.'">wróć</a>';
-        }
-               
-		                                 
-                                    
+    if (empty($_POST['dodaj_element1']) || empty($_POST['dodaj_element2']))
+    {
+        exit(header("Location: tryb_edycji.php?zestaw=$table"));
+    }                
+                                                
+    //if ($_POST['dodaj_element1'] == $dodaj_element1  && $_POST['dodaj_element2'] == $dodaj_element2 && $_POST['dodaj_element3'] == $dodaj_element3) 
+    //{
+    //exit(header("Location: tryb_edycji.php?zestaw=$tabela"));
+    //}
+    
+    try
+    {
+         $polecenie = "INSERT INTO $tabela ($angielski, $polski, $przyklad , $zdanie, $flaga_baza ) VALUES ('$dodaj_element1', '$dodaj_element2', '1', '$dodaj_element3', '0')";
+         $liczba = $pdo ->exec($polecenie);
+         exit(header("Location: tryb_edycji.php?zestaw=$table"));
+    }
+    catch(PDOException $e)
+    {
+       echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
+       echo '</br><a href="tryb_edycji.php?zestaw='.$table.'">wróć</a>';
+    }
+
 break;
                                   
 case 'edytuj':
@@ -273,31 +174,31 @@ $slowo = filtruj($_POST['slowo']);
       //header("Location: tryb_edycji.php?zestaw=$tabela");
     }
 
-  	  try
-      {
-      $zapytanie = "SELECT * FROM $tabela WHERE $angielski LIKE '$slowo' OR $polski LIKE '$slowo'";
-      echo $zapytanie;              
-      $liczba = $pdo ->query($zapytanie) or die('Błąd zapytania');
-      $wykonanie = $pdo->prepare($zapytanie);
-      $wykonanie->execute();
-      $licznik_id=$wykonanie->rowCount();
-         if ($licznik_id == 0)
-         { 
-                header("Location: tryb_edycji.php?zestaw=$tabela");
-         }
-         else 
-         {
-                while($row = $liczba->fetch())
-                              {                 
-                              wyswietlam($tabela, $row);
-                              }
-         }                                                          
-      }
-      catch(PDOException $e)
-      {
-          echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
-          echo '</br><a href="tryb_edycji.php?zestaw='.$tabela.'">wróć</a>';
-      }
+    try
+    {
+    $zapytanie = "SELECT * FROM $tabela WHERE $angielski LIKE '$slowo' OR $polski LIKE '$slowo'";
+    echo $zapytanie;              
+    $liczba = $pdo ->query($zapytanie) or die('Błąd zapytania');
+    $wykonanie = $pdo->prepare($zapytanie);
+    $wykonanie->execute();
+    $licznik_id=$wykonanie->rowCount();
+        if ($licznik_id == 0)
+        { 
+            header("Location: tryb_edycji.php?zestaw=$tabela");
+        }
+        else 
+        {
+            while($row = $liczba->fetch())
+                            {                 
+                            load_word_by_id($tabela, $row);
+                            }
+        }                                                          
+    }
+    catch(PDOException $e)
+    {
+        echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
+        echo '</br><a href="tryb_edycji.php?zestaw='.$tabela.'">wróć</a>';
+    }
 
 break;
 
@@ -345,21 +246,7 @@ break;
                               
 default:
 
-
-    try
-    {
-        $liczba = $pdo -> query("SELECT * FROM $tabela"); 
-                                          
-            while($row = $liczba->fetch())
-            {
-                wyswietlam($tabela, $row);
-            }
-                  
-    }
-    catch(PDOException $e)
-    {
-        echo 'Połączenie nie mogło zostać utworzone: ' . $e->getMessage();
-    }
+load_fishcards_all($tabela, $pdo);
     
 break;
 
@@ -367,6 +254,10 @@ break;
 
 ?>
 </body>
+<script src="js/activate_the_word_set.js"></script>
+<script src="js/deactivate_the_word_set.js"></script>
+<script src="js/routes.js"></script>
+
 
 <script>
 
@@ -374,50 +265,17 @@ const menu_click = document.querySelector('.menubutton').addEventListener('click
     document.querySelector('.menu').classList.toggle('active');
 });
 
-/* skrypty odpowiedzialne za przemieszczanie się między stronami*/
 
-  $('#przycisk').click(function(){
-    window.document.location.href="fiszki.php";
-    return false;
-  });
-
-
-
-  $('#zestawy').click(function(){
-    window.document.location.href="tryb_wyboru.php";
-    return false;
-  });
-
-
-$(document).ready(function(){
-  $('#aktywuj').click(function(){
-    let queryString = window.location.search;
-    let urlParams = new URLSearchParams(queryString);
-    let zestaw = urlParams.get('zestaw');
-    window.document.location.href="tryb_edycji.php?zestaw="+zestaw+"&"+zestaw+"=aktywuj";
-    return false;
-  });
-});
-
-$(document).ready(function(){
-  $('#deaktywuj').click(function(){
-    let queryString = window.location.search;
-    let urlParams = new URLSearchParams(queryString);
-    let zestaw = urlParams.get('zestaw');
-    window.document.location.href="tryb_edycji.php?zestaw="+zestaw+"&"+zestaw+"=deaktywuj";
-    return false;
-  });
-});
 
 
 //----------Funkcje obslugujące zestaw - jeszcze nie działają i powodują błędy ---------//
 
-    $(document).ready(function() {
-        $('.wybrany_zestaw').click(function(){
-            $('.wybrany_zestaw').removeClass('active')
-            console.log(this.classList.add('active'))
-        })
-   })
+$(document).ready(function() {
+    $('.wybrany_zestaw').click(function(){
+        $('.wybrany_zestaw').removeClass('active')
+        console.log(this.classList.add('active'))
+    })
+})
 
 // usun_zestaw
 </script>
